feature: Create starter files to integrate Riot API
