+++ backend/.env.example
@@ -2,3 +2,4 @@ DB_USERNAME=
 DB_PASSWORD=
 RIOT_API_KEY=
 PORT=
+SEED_PASSWORD=
