+++ backend/package.json
@@ -35,6 +35,7 @@
     "typescript": "^4.8.4"
   },
   "dependencies": {
+    "axios": "^1.1.3",
     "dotenv": "^16.0.3",
     "express": "^4.18.1",
     "express-status-monitor": "^1.3.4",
